requirements.txt
flask
requests
openai
python-telegram-bot==13.15
